requirements.txt
# python 3.11.2 used
bs4==0.0.2
pandas==2.2.2
PyYAML==6.0.1
selenium==4.18.1
